Protect edit and delete product view

+++ marketplace/myapp/views.py
@@ -98,6 +98,9 @@ def create_product(request):
 
 def edit_product(request, id):
     product = Product.objects.get(id=id)
+    if product.seller != request.user:
+        return redirect('invalid')
+
     product_form = ProductForm(request.POST or None, request.FILES or None, instance=product)
     if request.method == "POST" and product_form.is_valid():
         product_form.save()
@@ -107,6 +110,8 @@ def edit_product(request, id):
 
 def delete_product(request, id):
     product = Product.objects.get(id=id)
+    if product.seller != request.user:
+        return redirect('invalid')
     if request.method == "POST":
         product.delete()
         return redirect('index')
@@ -129,3 +134,7 @@ def register(request):
 
     form = UserRegistrationForm()
     return render(request, 'myapp/register.html', {'form': form})
+
+
+def invalid(request):
+    return render(request, 'myapp/invalid.html')
